added scene system

+++ GLTFLoader.cpp
@@ -39,11 +39,18 @@ bool GLTFLoader::LoadModel(const std::string& path) {
         }
         else {
             glm::vec3 translation = node.translation.size() == 3 ?
-                glm::vec3(node.translation[0], node.translation[1], node.translation[2]) : glm::vec3(0.0f);
+                glm::vec3(static_cast<float>(node.translation[0]),
+                    static_cast<float>(node.translation[1]),
+                    static_cast<float>(node.translation[2])) : glm::vec3(0.0f);
             glm::vec3 scale = node.scale.size() == 3 ?
-                glm::vec3(node.scale[0], node.scale[1], node.scale[2]) : glm::vec3(1.0f);
+                glm::vec3(static_cast<float>(node.scale[0]),
+                    static_cast<float>(node.scale[1]),
+                    static_cast<float>(node.scale[2])) : glm::vec3(1.0f);
             glm::quat rotation = node.rotation.size() == 4 ?
-                glm::quat(node.rotation[3], node.rotation[0], node.rotation[1], node.rotation[2]) : glm::quat();
+                glm::quat(static_cast<float>(node.rotation[3]),
+                    static_cast<float>(node.rotation[0]),
+                    static_cast<float>(node.rotation[1]),
+                    static_cast<float>(node.rotation[2])) : glm::quat();
 
             transform = glm::translate(glm::mat4(1.0f), translation) *
                 glm::toMat4(rotation) *
